Fix: properly convert phase name to CString and actlog sleep 100ms (not s); shared code for actmon and actlog
